backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
python-multipart==0.0.9
httpx==0.27.0

# APIs
firecrawl==1.15.0
elevenlabs==1.50.0
google-genai==1.14.0

# Database
supabase==2.11.0

# Audio processing
pydub==0.25.1

# Async tasks
celery[redis]==5.4.0

# Utilities
pydantic==2.9.0
